RSC convert admin receptionists page
